Testing

+++ ExaminationSystem/Features/StudentDashboard/StudentDashboardResponseDto.cs
@@ -30,7 +30,7 @@ public class RecentQuizAttemptDto
 
     public string QuizTitle { get; init; } = null!;
 
-    public int Score { get; init; }
+    public double Score { get; init; }
 
     public bool Passed { get; init; }
